### 2024-03-22 #6521 @david-mathews-1994
Apologies for the different paper format than expected. JOSS appears to expect extremely different things from most traditional journals, and this paper was originally intended for a different destination so its format is more aligned with those requirements. I'll work on tweaking it to be more in line with the expectation of this journal.
