### 2024-03-22 7c125ce
Update paper.md
initial cutting of content to match JOSS requirements
