+++ paper/paper.md
@@ -106,25 +106,21 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 
 # Summary
 
-Delta-Rice, an efficient method for high performance lossless compression of digitized analog signals within the HDF5 library and data format, is presented. This algorithm was developed for the Nab experiment, a fundamental neutron physics experiment, but is broadly applicable to other experimental efforts requiring lossless compression of noisy digitized detector data. This is a two-step process where incoming data is passed through a pre-processing filter, by default delta encoding, to prepare the data for a traditional Rice compression routine. A routine for determining the optimal pre-processing filter for a dataset is demonstrated. Also demonstrated are modifications to support deployment of this algorithm to both FPGA and GPU architectures. When applied to data collected by the Nab data acquisition system, this method produced output files 29% their initial size, and was able to do so with an average read/write throughput in excess of 2 GB/s on a single CPU. Compared to the widely used G-zip compression routine, Delta-Rice reduces the file size by 33% more with over an order of magnitude increase in read/write throughput. Delta-Rice is available on CPU to users through the HDF5 library.[^1]
+Delta-Rice, an efficient method for high performance lossless compression of digitized analog signals within the HDF5 library[@hdf5] and data format, is presented. This algorithm was developed for the Nab experiment, a fundamental neutron physics experiment, but is broadly applicable to other experimental efforts requiring lossless compression of noisy digitized detector data. This is a two-step process where incoming data is passed through a pre-processing filter, by default delta encoding, to prepare the data for a traditional Rice compression routine. A routine for determining the optimal pre-processing filter for a dataset is demonstrated. Also demonstrated are modifications to support deployment of this algorithm to both FPGA and GPU architectures. When applied to data collected by the Nab data acquisition system, this method produced output files 29% their initial size, and was able to do so with an average read/write throughput in excess of 2 GB/s on a single CPU. Compared to the widely used G-zip [@gzip] compression routine, Delta-Rice reduces the file size by 33% more with over an order of magnitude increase in read/write throughput. Delta-Rice is available on CPU to users through the HDF5 library.[^1]
 
-# Introduction
+# Statement of Need
 
-The ability to digitize, process, and record analog signals is a critical part of many modern physics experiments. For large-scale or high-statistics experiments in particular, the resulting datasets can require petabytes of storage leading to difficulties in both storage and data distribution. The Nab experiment, which will precisely characterize the decay of free neutrons, is one such high-statistics experiment [@Fry2019]. During the multi-year experiment, $10^9$ events will be measured with a data acquisition (DAQ) system recording between $100$ and $400$ MB/s of digitized detector data [@david_2022]. The resulting dataset is expected to reach several petabytes in size. The cost and complexity of storing and distributing this dataset, along with the requirements described below, motivated the development of a custom compression routine for the 1-dimensional time-series detector data collected by the DAQ that represents over 95% of the output size. 
-
[…]
