### 2024-03-22 e3d321e
Update paper.md
shortening it even further
