### 2024-03-22 3c535e5
Update paper.md
removing more
