### 2024-03-22 069d5e5
Update paper.md
removing FPGA references
