### 2024-03-22 2ad34f6
Update paper.bib
adding cupy reference
