+++ paper/paper.bib
@@ -229,3 +229,11 @@ title = {A new cryogenic apparatus to search for the neutron electric dipole mom
 journal = {Journal of Instrumentation},
 abstract = {A cryogenic apparatus is described that enables a new experiment, nEDM@SNS, with a major improvement in sensitivity compared to the existing limit in the search for a neutron Electric Dipole Moment (EDM). This apparatus uses superfluid 4He to produce a high density of Ultra-Cold Neutrons (UCN) which are contained in a suitably coated pair of measurement cells. The experiment, to be operated at the Spallation Neutron Source at Oak Ridge National Laboratory, uses polarized 3He from an Atomic Beam Source injected into the superfluid 4He and transported to the measurement cells where it serves as a co-magnetometer. The superfluid 4He is also used as an insulating medium allowing significantly higher electric fields, compared to previous experiments, to be maintained across the measurement cells. These features provide an ultimate statistical uncertainty for the EDM of 2−3× 10−28 e-cm, with anticipated systematic uncertainties below this level.}
 }
+
+@inproceedings{cupy,
+  author       = "Okuta, Ryosuke and Unno, Yuya and Nishino, Daisuke and Hido, Shohei and Loomis, Crissman",
+  title        = "CuPy: A NumPy-Compatible Library for NVIDIA GPU Calculations",
+  booktitle    = "Proceedings of Workshop on Machine Learning Systems (LearningSys) in The Thirty-first Annual Conference on Neural Information Processing Systems (NIPS)",
+  year         = "2017",
+  url          = "http://learningsys.org/nips17/assets/papers/paper_16.pdf"
+}
