### 2024-03-22 #6521 @david-mathews-1994
@danielskatz It appears the word count from the robot is also including tables, author information, as well as acknowledgements, but not references. Just the author list and funding acknowledgements is 565 words. Are these words that are taken into account when trying to keep the papers between 250 and 1000 words? For example, if I only count the words of actual paper content (section labels, each paragraph, image captions), ignoring tables, figures, author information, references, and acknowledgements, I'm seeing around 1800 words. I just want to be sure I'm understanding the recommendation when I'm doing my editing.
