### 2024-03-22 #6521 @danielskatz
@david-mathews-1994 - I think you may be right about how the word count is done.  However, having said that, your paper still seems quite long to me. I think if you could get your current self-measurement of 1800 words to 1000, this would likely be good. I think much of the performance section could be a performance.md file in the repo, with a small bit kept in the paper and then a link to that file, for example.
