### 2024-03-22 #6521 @david-mathews-1994
@danielskatz It's definitely a little bit too long still, I can work on that. I like the suggestion about the performance information. I'll see about moving that and some of the figures to a different document to keep it neater. I'm used to writing documents like this in Latex and not markdown files so the formatting of the various figures and tables is a bit rough at the moment. I appreciate the advice and I'll keep working on this.
