### 2024-03-25 17b7be9
Update Algorithm.md
updating the algorithm to adjust the images
