### 2024-03-25 cc2f758
Update Algorithm.md
adjusting image references again
