+++ docs/Algorithm.md
@@ -19,7 +19,7 @@ This compression routine can be reconstructed by parsing each bit within the 32 
 
 Optimization of both $m$ and $c$ requires knowledge of the probability density function of values after preparatory encoding, $P(x)$. Determining of $P(x)$ can be done either on a per-file basis which will provide the best possible compression from this method at the expense of throughput, or the assumption can be made that $P(x)$ does not change significantly from file to file in a dataset and can be determined once for a subset of the data. The optimal values for $m$ and $c$ are found via minimization of $$ B(m, c) = \sum_i P(x_i) * b(x_i, m, c, b_0) $$ where $b(x_i, m, c, b_0)$ is the number of bits required for each datapoint and $b_0$ is the number of bits used in the original representation of each value $x_i$. The sum is over all possible values of $x_i$ as defined by the finite range of digitized values. For example, $x_i$ ranges from $0$ to $16383$ for 14 bit digitizers. For the case of a Gaussian probability distribution $P(x)$ centered about $0$, the relationship between the standard deviation $\sigma$ and $m$ is explored in Figure \ref{fig:CorrelationPlot}. Note that datasets with smaller standard deviations will compress better than larger standard deviation datasets, but are also more sensitive to the choice of $m$ parameter. In this work, the cutoff value $c$ is fixed to 8, but in general the cutoff value $c$ should be optimized in tandem with $m$. 
 
-![Compression Ratio as a function of $m$ and $\sigma$ for Gaussian distributions using Rice Coding. Locations marked with $\star$ represent the best compression for that $\sigma$ value. No cutoff value $c$ was used during this optimization to prevent the representation of $q$ from expanding past the initial storage size. The values of $x$ ranged from $-8192$ to $8192$ as if from a signed 14 bit digitizer](images/CompressionRatioPlot.png)
+![Compression Ratio as a function of $m$ and $\sigma$ for Gaussian distributions using Rice Coding. Locations marked with $\star$ represent the best compression for that $\sigma$ value. No cutoff value $c$ was used during this optimization to prevent the representation of $q$ from expanding past the initial storage size. The values of $x$ ranged from $-8192$ to $8192$ as if from a signed 14 bit digitizer](./images/CompressionRatioPlot.png?raw=true)
 
 ## Preparatory Encoding
 The goal of this encoding operation is to manipulate the incoming data into a form that is more optimal for Rice Coding. As shown for a Gaussian probability distribution in Figure \ref{fig:CorrelationPlot}, a dataset with a smaller standard deviation can be compressed to a smaller overall size compared to a dataset with a larger standard deviation. This makes intuitive sense as a dataset with a smaller distribution centered around $0$ will have a smaller average value $q$ for a given $m$, thus reducing the output size of the compressed data. In addition to reducing the spread in the data and centering the values around $0$, any encoding operation must be both reversible, so the encoding can be undone when the data is de-compressed, and maintain integer precision, as Rice encoding is defined for integer input values $x$. 
@@ -28,4 +28,4 @@ The default encoding operation performed in this routine is delta encoding. This
 
 Delta encoding may not be the ideal encoding operation for a given dataset so this compression routine supports the use of alternative encoding filters. These alternate filters are implemented via convolution with the original data, which can be performed iteratively or recursively depending on the hardware. These filters are preferably small as the filter must be stored alongside the compressed data to allow for decompression. The functions for encoding and decoding data with an arbitrary filter can be seen in the source code. Delta encoding can also be expressed in terms of a convolutional filter, in which case it is defined as $[1, -1]$. Determination of the optimal encoding filter is discussed in the Optimization.md file within the docs/ folder in the repository.
 
[…]
