### 2024-03-25 3704f93
Update Optimization.md
updating the optimization document with the encoding filter discussion
