### 2024-03-25 022400a
Update Optimization.md
adding in M optimization discussion and image
