### 2024-03-25 2da8da0
Update Optimization.md
trying to get the image linking to work
