### 2024-03-25 11c1fdc
Update paper.md
Moved most content to the other documents
