### 2024-03-25 13ff52d
Update paper.md
adding the rice coding section label back in
