### 2024-03-25 083e2b8
Update paper.md
Updating image reference
