### 2024-03-25 2316e9f
Create Performance.md
adding in the performance documentation
