+++ docs/Performance.md
@@ -1,0 +1,92 @@
+# Performance
+
+The performance of the algorithm was tested against several other compression routines to determine which is best suited for deployment in real-time data acquisition and analysis systems. Both the reduction in size and the rate of the compression were compared. All tests, unless otherwise stated, were performed on the machine described here. A RAM disk was used for the storage allocation to allow for peak throughput and reduce the effect of disk I/O on the results. 
+
+ - OS: Ubuntu 22.04
+ - CPU: AMD Threadripper Pro 5955WX
+ - RAM: 4x32 GB DDR4 @ 3200 MHz
+ - Storage: tmpfs
+
+The competing algorithms used during testing are all commonly available in HDF5 implementations: Gzip, LZF, and Bitshuffle. Testing was performed in Python using the h5py library to manage the reading and writing from the HDF5 file [@h5py]. Each test was performed 10 times to reduce random variance in the results and the mean of these tests is reported. For all testing, the initial data were stored in memory before being written to the file. All data were written to a ramdisk configured with tmpfs to reduce dependence on hard-drive read/write performance. An additional set of benchmarks were done without compression enabled to establish the baseline file I/O performance in the system for each dataset. The Delta-Rice algorithm was run in both single-threaded and multi-threaded mode to show the scaling performance of the algorithm and used the default $m=8$ parameter and the delta encoding filter for all datasets. Gzip, LZF, and Bitshuffle were all run with the default parameters. Each routine was run with the same chunksize, the dimensions of the uncompressed data in memory, to keep the chunking behavior consistent. Recorded signals from the Nab and NOPTREX experiments were used to test this routine, along with simulated detector signals from the nEDM@SNS experiment [@nedmsns]. The figure below shows examples of these signals. Each signal type has very different shape features and lengths deliberately to test the flexibility of the algorithms. 
+
+![](./images/ComparingSignalFeatures.png?raw=true)
+*Comparing the features of the different signal types from Nab, nEDM@SNS, and NOPTREX. Each dataset has significantly different signal shapes and features. Both Nab and NOPTREX share exponentially decaying signal shapes, but with opposite polarities and different noise profiles while nEDM@SNS is a square pulse.*
+
+The first results are for the Nab experiment dataset. All methods used the same chunksize of $2000\times7000$ and the segment length in Delta-Rice was set to 7000. 
+
+| Compression Method      | File Size (%) | Read (MB/s) | Write (MB/s) |
+|-------------------------|---------------|-------------|--------------|
+| No Compression          | 100           | 3587        | 2906         |
+| Gzip                    | 39            | 374         | 71           |
+| LZF                     | 59            | 717         | 554          |
+| Bitshuffle (32 threads) | 38            | 1852        | 2131         |
+| Delta-Rice (1-thread)   | 29            | 229         | 500          |
+| Delta-Rice (32-thread)  | 29            | 1782        | 2387         |
+
+The second set of tests was run on simulated signals from the nEDM@SNS experiment. All methods used a chunksize of $32\times81920$ and a segment length of $81920$ was set for the Delta-Rice method. 
+
+| Compression Method      | File Size (%) | Read (MB/s) | Write (MB/s) |
+|-------------------------|---------------|-------------|--------------|
+| No Compression          | 100           | 3759        | 3799         |
+| Gzip                    | 29            | 440         | 75           |
+| LZF                     | 49            | 722         | 662          |
+| Bitshuffle (32 threads) | 30            | 1833        | 1353         |
+| Delta-Rice (1-thread)   | 27            | 436         | 539          |
+| Delta-Rice (32-thread)  | 27            | 1717        | 2084         |
+
+The third and final set of testing was performed on data from the NOPTREX collaboration. All methods used a chunksize of $32\times500000$ with a segment length of $500000$ specified for the Delta-Rice method. 
+
[…]
