+++ docs/Performance.md
@@ -10,6 +10,7 @@ The performance of the algorithm was tested against several other compression ro
 The competing algorithms used during testing are all commonly available in HDF5 implementations: Gzip, LZF, and Bitshuffle. Testing was performed in Python using the h5py library to manage the reading and writing from the HDF5 file [@h5py]. Each test was performed 10 times to reduce random variance in the results and the mean of these tests is reported. For all testing, the initial data were stored in memory before being written to the file. All data were written to a ramdisk configured with tmpfs to reduce dependence on hard-drive read/write performance. An additional set of benchmarks were done without compression enabled to establish the baseline file I/O performance in the system for each dataset. The Delta-Rice algorithm was run in both single-threaded and multi-threaded mode to show the scaling performance of the algorithm and used the default $m=8$ parameter and the delta encoding filter for all datasets. Gzip, LZF, and Bitshuffle were all run with the default parameters. Each routine was run with the same chunksize, the dimensions of the uncompressed data in memory, to keep the chunking behavior consistent. Recorded signals from the Nab and NOPTREX experiments were used to test this routine, along with simulated detector signals from the nEDM@SNS experiment [@nedmsns]. The figure below shows examples of these signals. Each signal type has very different shape features and lengths deliberately to test the flexibility of the algorithms. 
 
 ![](./images/ComparingSignalFeatures.png?raw=true)
+
 *Comparing the features of the different signal types from Nab, nEDM@SNS, and NOPTREX. Each dataset has significantly different signal shapes and features. Both Nab and NOPTREX share exponentially decaying signal shapes, but with opposite polarities and different noise profiles while nEDM@SNS is a square pulse.*
 
 The first results are for the Nab experiment dataset. All methods used the same chunksize of $2000\times7000$ and the segment length in Delta-Rice was set to 7000. 
