### 2024-03-25 4996a5f
Update paper.md
changing some of the links to be global not locally referencing
