+++ paper/paper.md
@@ -118,7 +118,7 @@ This algorithm is a two-step process: the digitized signal is first passed throu
 ## Rice Coding 
 Rice coding functions by encoding a value $x$ in 2 pieces: $q$, the result of a division by a tunable parameter $m$, and $r$, the remainder of that division. $q$ is stored in Unary coding, with $r$ in truncated binary. In this routine, signed values are handled by interleaving positive and negative values as follows: $x\prime = 2*x$ for $x>=0$ and $x\prime = 2|x|-1$ for $x<0$. This routines uses Rice coding instead of the more general Golomb coding [@golomb] because the restriction to powers of $2$ for $m$ allows for more efficient calculations. In the case that $q>=8$, this routine will output $q=8$ followed by the original number in 16-bit signed representation. This is done to ensure that the amount a value can fail to be compressed is fixed. It is the responsibility of the preparatory encoding operation to reduce the chance of this occuring. The value $m$ is stored alongside the compressed data for decompression. 
 
-The outputs from this method are packed sequentially into a temporary 64 bit storage buffer. Once more than 32 bits of this buffer are used, the first 32 bits are output to the file and the remaining bits are shifted over to the start and the process continues. A simple example for a 16 bit storage buffer with 8 bit output is shown here for storing $-2$ and $25$ with $m=8$. For information about the optimization of $m$, see [Optimization](./../docs/Optimization.md). 
+The outputs from this method are packed sequentially into a temporary 64 bit storage buffer. Once more than 32 bits of this buffer are used, the first 32 bits are output to the file and the remaining bits are shifted over to the start and the process continues. A simple example for a 16 bit storage buffer with 8 bit output is shown here for storing $-2$ and $25$ with $m=8$. For information about the optimization of $m$, see [Optimization](https://github.com/david-mathews-1994/deltarice/blob/master/docs/Optimization.md). 
 
 | Bit          | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14 | 15 |
 |--------------|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|---|
@@ -128,7 +128,7 @@ The outputs from this method are packed sequentially into a temporary 64 bit sto
 | After Output | 0 | 0 | 1 | 0 | 1 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 |
 
 ## Preparatory Encoding
-Preparatory encoding is done to adjust the dataset to a form more optimal for Rice Coding. By default, this is done with delta encoding which stores the difference of subsequent values. The image below shows an example of this when applied to a signal from the Nab experiment. A simple optimization routine for determining the ideal filter is discussed in [Optimization](./../docs/Optimization.md).
+Preparatory encoding is done to adjust the dataset to a form more optimal for Rice Coding. By default, this is done with delta encoding which stores the difference of subsequent values. The image below shows an example of this when applied to a signal from the Nab experiment. A simple optimization routine for determining the ideal filter is discussed in [Optimization](https://github.com/david-mathews-1994/deltarice/blob/master/docs/Optimization.md). 
 
 ![](./../docs/images/ExampleEncoding.png?raw=true)
 
@@ -136,7 +136,7 @@ Preparatory encoding is done to adjust the dataset to a form more optimal for Ri
 
 # Implementation
 
[…]
