### 2024-03-25 a2f66eb
Update paper.md
removed conclusion section and re-worked the Rice Coding section a bit
