### 2024-03-25 #6521 @david-mathews-1994
@danielskatz I'm getting a word count of just below 1000 words when only counting the actual content (ignoring stuff like tables and author information). I went through and moved a lot of the specifics to new markdown files in the docs/ folder within the repo so that information is still available to the end user if they're interested.
