### 2024-03-25 #6521 @danielskatz
@david-mathews-1994 - please check to see if the one specific missing DOI above is correct (as it appears to be), and if so, please add it to your bib file, the regenerate your pdf again.
