### 2024-03-25 b09f872
Update paper.bib
updating doi information and removing some unused references
