+++ paper/paper.bib
@@ -5,7 +5,7 @@ volume = {969},
 pages = {163961},
 year = {2020},
 issn = {0168-9002},
-doi = {https://doi.org/10.1016/j.nima.2020.163961},
+doi = {10.1016/j.nima.2020.163961},
 author = {D.C. Schaper and C. Auton and L. Barrón-Palos and M. Borrego and A. Chavez and L. Cole and C.B. Crawford and J. Curole and H. Dhahri and K.A. Dickerson and J. Doskow and W. Fox and M.H. Gervais and B.M. Goodson and K. Knickerbocker and C. Jiang and P.M. King and H. Lu and M. Mocko and D. Olivera-Velarde and J.G. {Otero Munoz} and S.I. Penttilä and A. Pérez-Martín and B. Short and W.M. Snow and K. Steffen and J. Vanderwerp and G. Visser},
 keywords = {Parity violation, Scintillator detector, Neutron resonance, Fundamental symmetry, Time reversal violation},
 abstract = {We describe a modular apparatus for use in parity-violation measurements in epithermal neutron–nucleus resonances with high instantaneous neutron fluxes at the Manuel Lujan Jr. Neutron Scattering Center at Los Alamos National Laboratory. This apparatus is designed to conduct high-precision measurements of the parity-odd transmission asymmetry of longitudinally polarized neutrons through targets containing nuclei with p-wave neutron–nucleus resonances in the 0.1–10 eV energy regime and to accommodate a future search for time reversal violation in polarized neutron transmission through polarized nuclear targets. The apparatus consists of an adjustable neutron and gamma collimation system, a 3He-4He ion chamber neutron flux monitor, two identical cryostats for target cooling, an adiabatic eV-neutron spin flipper, a near-unit efficiency 6Li-7Li scintillation detector operated in current mode, a flexible CAEN data acquisition system, and a neutron spin filter based on spin-exchange optical pumping of 3He gas. We describe the features of the apparatus design devoted to the suppression of systematic errors in parity-odd asymmetry measurements. We describe the configuration of the apparatus used to conduct a precision measurement of parity violation at the 0.7 eV p-wave resonance in 139La which employs two identical 139La targets, one to polarize the beam on the p-wave resonance using the weak interaction and one to analyze the polarization.}
@@ -55,13 +55,6 @@ abstract = {We describe a modular apparatus for use in parity-violation measurem
     doi = {10.1007/s10751-018-1538-7}
 }
 
-@phdthesis{calcium-45, 
-    title={Measurement of the Fierz Interference Term for Calcium-45}, 
-    author={Birge, Noah}, 
-    year={2019},
-    school={University of Tennessee, Knoxville}
-}
-
 @article{MASUI2015181,
     title = {A compression scheme for radio data in high performance computing},
     journal = {Astronomy and Computing},
@@ -69,51 +62,34 @@ abstract = {We describe a modular apparatus for use in parity-violation measurem
     pages = {181-190},
     year = {2015},
     issn = {2213-1337},
-    doi = {https://doi.org/10.1016/j.ascom.2015.07.002},
+    doi = {10.1016/j.ascom.2015.07.002},
     
     author = {K. Masui and M. Amiri and L. Connor and M. Deng and M. Fandino and C. Höfer and M. Halpern and D. Hanna and A.D. Hincks and G. Hinshaw and J.M. Parra and L.B. Newburgh and J.R. Shaw and K. Vanderlinde},
     keywords = {Radio astronomy, Data compression, HDF5, High performance computing},
[…]
