+++ paper/paper.bib
@@ -179,8 +179,8 @@ note =
     Publisher = {IEEE},
     Title = {OpenMP: an industry standard API for shared-memory programming},
     Volume = {5},
-    Year = {1998}
-    doi={10.1109/99.660313}
+    Year = {1998},
+    doi={10.1109/99.660313},
 }
 
 @article{nedmsns,
