### 2024-03-25 d7d76ab
Update paper.md
removing gzip reference
