### 2024-03-25 44667a5
Update paper.bib
removed blank note line that may have broken joss bot
