+++ paper/paper.bib
@@ -75,9 +75,9 @@ author = {The HDF Group, None and Koziol, Quincey and USDOE Office of Science},
 abstractNote = {The HDF5 library and file format implement the HDF5 data model for storing and managing data. This implementation supports an unlimited variety of datatypes and is designed for flexible and efficient I/O and for high volume and complex data. It is highly portable and extensible, allowing applications to evolve in their use of HDF5. The HDF5 Technology suite includes tools and applications for managing, manipulating, viewing, and analyzing data in the HDF5 format and other implementations.},
 url = {https://www.osti.gov//servlets/purl/1631295},
 doi = {10.11578/dc.20180330.1},
-url = {https://www.osti.gov/biblio/1631295}, year = {2020},
+url = {https://www.osti.gov/biblio/1631295}, 
+year = {2020},
 month = {2},
-note =
 } 
 
 @phdthesis{david_2022,
