### 2024-04-05 #6521 @samcunliffe
Hello @emdupre

Thanks for the ping! I'm afraid for the next ~2 months I don't have the capacity to take this on. This probably doesn't help you, but I would have capacity _in_ 2 months (i.e. June 2024++).

For what it's worth the paper looks very interesting! (and in my domain etc)
Sam.
