### 2024-04-08 #6521 @emdupre
👋 Hi  @DanNixon,

Would you be willing to review _Delta-Rice: A HDF5 Compression Plugin optimized for Digitized Detector Data_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/david-mathews-1994/deltarice) and this short paper: https://github.com/openjournals/joss-papers/blob/joss.06521/joss.06521/10.21105.joss.06521.pdf

If you're available to review this work within the next month—or six weeks, at most—please let me know and I'll add you as a reviewer.

Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
