### 2024-04-09 #6598 @editorialbot
[REVIEW]: Delta-Rice: A HDF5 Compression Plugin optimized for Digitized Detector Data
**Submitting author:** @david-mathews-1994 (<a href="http://orcid.org/0000-0002-4897-4379">David Mathews</a>)
**Repository:** https://github.com/david-mathews-1994/deltarice
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
