### 2024-04-23 #6598 @macrocosme
There doesn't seem to be community guidelines. It would be good to add a note. Tests are note specifically provided, but the examples can serve as to define some, so OK for now.
