### 2024-04-25 f893d1d
fixed typo in paper.md
