### 2024-04-25 #6598 @dineshchitlangia
@emdupre I have finished my review and updated the checklist.

Overall, the body of work looks good to me.
- I opened a PR to fix a typo, hope it can be merged by the author.
- It will be great to add a CONTRIBUTING.md that describes the community guidelines for questions/issues/feature requests and contributions.
- Recommended to add multiple test cases covering some positive and negative test case scenarios and edge cases if any. Such test cases not only ensure the correctness but it also make it more reliable and hence easy to trust/adopt.

Thank you for the invitation to review.
