### 2024-04-26 david-mathews-1994/deltarice#3 @DanNixon
CI/automated testing
Ideally it would be good to see compilation and some form of round trip testing (even if relatively minimal) done automatically.
