### 2024-04-26 david-mathews-1994/deltarice#4 @DanNixon
Spurious files in repo
There appear to be a number of files/directories in the repo that are seemingly not there for a reason or should have been in `.gitignore`:

- `.eggs`
- `.ipynb_checkpoints`
- `conda-recipe/.ipynb_checkpoints`
- `src/.ipynb_checkpoints`
- `image.png`
- `image-1.png`
- `image-2.png`
