### 2024-04-26 #6598 @DanNixon
I've now finished my review, overall the work looks really good to me.

My only points were:

- Echoing what has already been said about contributing guides
- I have not ticked the testing point on my checklist, as there is no automated compilation or testing. The examples code does do a round trip, but this could easily be automated with GitHub Actions. (https://github.com/david-mathews-1994/deltarice/issues/3)
- A minor point, but there were a few random files in the repo that looked like they should not be there. (https://github.com/david-mathews-1994/deltarice/issues/4)
