### 2024-04-30 #6598 @emdupre
Thank you for confirming your reviews, @dineshchitlangia @macrocosme and @DanNixon ! And thank you for your responsiveness to date, @david-mathews-1994.

To follow-up on this question :
> I'd be happy to add some guidelines in a CONTRIBUTING.md file. Do you have any particular examples that I should follow?

We don't have a specific template that we request, and our [accepted papers](https://joss.theoj.org/papers/published) adopt a range of different formats. In general, though, I commonly see projects create a CONTRIBUTING.md ([as GitHub recognizes these files](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)). [Here is one example CONTRIBUTING.md file](https://github.com/KennethEnevoldsen/augmenty/blob/main/CONTRIBUTING.md) from a project recently published in JOSS.

If you have any other questions, please don't hesitate to let me know !
