### 2024-05-01 c58a02a
Create CONTRIBUTING.md
Added a contributing guide
