### 2024-05-01 #6598 @david-mathews-1994
@emdupre I added in a contribution guide to the repo and cleaned up the files that @DanNixon identified. I'll look into an automated test suite for the repository as well, and will probably expand the testing to be a bit more thorough than what is presently defined in the example scripts (which I will probably try to use in the automated testing).

What is the path forward for this submission at this stage? This is my first JOSS submission so I'm not familiar with what the next steps are.
