### 2024-05-01 david-mathews-1994/deltarice#4 @david-mathews-1994
Thank you for flagging these files! They should be all removed now and I've added a few additional lines to the .gitignore file to hopefully prevent this in the future
