### 2024-05-01 #6598 @emdupre
> What is the path forward for this submission at this stage? This is my first JOSS submission so I'm not familiar with what the next steps are.

At this point, we ask that you address all of the comments / issues that the authors have raised and then respond in this thread with any remaining questions or blockers. This will give the reviewers a chance to see if your changes have addressed their concerns !

Specifically, given @dineshchitlangia 's comment :
> Recommended to add multiple test cases covering some positive and negative test case scenarios and edge cases if any. Such test cases not only ensure the correctness but it also make it more reliable and hence easy to trust/adopt.

and @DanNixon 's issue : https://github.com/david-mathews-1994/deltarice/issues/3

It seems that the remaining points are on developing a more extensive -- and automatic -- test suite.

Please let me know if this is clear, or if you have any other questions re : process that I can help with !
