### 2024-05-03 david-mathews-1994/deltarice#3 @david-mathews-1994
I'm working on a new set of tests in this branch: https://github.com/david-mathews-1994/deltarice/tree/tests. Mostly I'm working on making the C++ and Python test scripts more robust (Python is done, working on C++), and once those are fully testing the library properly I'll work on automating the routine.
