### 2024-05-08 7a871ce
First attempt at doing the CI work
