### 2024-05-08 7c6bf0b
Update main.yml
removing the 3.6 and 3.7 python versions since they seem to fail due to missing support for x64
