### 2024-05-08 david-mathews-1994/deltarice#3 @david-mathews-1994
In theory this has been resolved! I did a bit of testing and like you said, it wasn't super hard to set up. It runs a series of tests on Ubuntu and MacOS and goes through and runs the pythonTesting.py script to ensure it properly behaves. That script also has been modified to make sure that all of the possible input values are treated properly.
