### 2024-05-08 #6598 @david-mathews-1994
@emdupre Hello again! I think I have created the automated test suite. I did some testing on a different branch and it seemed to work just fine. Just pushed those updates to main and they're currently being tested. I will say it's a bit slow, but it does seem to be running properly at least which is good. Effectively it's checking the compilation and the proper running of the pythonExample.py script.
