### 2024-05-09 e562985
adding more verbose testing
