### 2024-05-09 a80f7b9
added more robust testing and patched a bug with invalid handling of different filters
