+++ pythonExample.py
@@ -4,30 +4,68 @@ import numpy as np
 import time
 
 
-#create test datasets that span the full range of 16-bit numbers
-dset1 = np.arange(-32768, 32768, dtype = np.int16)
-dset2 = np.arange(0, 2**16, dtype = np.uint16)
-bulkDset = np.zeros(shape = (100, dset2.shape[0]))
-
-bulkOptions = (8, bulkDset.shape[0])
-
+#worst case scenario testing with a bunch of random values
 f = h5py.File('testFile.h5', 'w')
-f.create_dataset('signed', data = dset1, compression=deltaRice.h5.H5FILTER)
-f.create_dataset('unsigned', data = dset2, compression=deltaRice.h5.H5FILTER)
-f.create_dataset('bulkDset', data = bulkDset, compression=deltaRice.h5.H5FILTER, compression_opts = bulkOptions)
-
+dset = np.random.uniform(-32768, 32768, size=2**16).astype(np.int16)
+f.create_dataset('test', data = dset, compression=deltaRice.h5.H5FILTER)
 f.close()
 
 f = h5py.File('testFile.h5', 'r')
-test1 = np.array_equal(f['signed'][()], dset1)
-test2 = np.array_equal(f['unsigned'][()], dset2)
-test3 = np.array_equal(f['bulkDset'][()], bulkDset)
+test = np.array_equal(f['test'][()], dset)
+f.close()
+assert test, "Failed default random number test"
+
+f = h5py.File('testFile.h5', 'w')
+compression_opts = (16,)
+f.create_dataset('test', data = dset, compression=deltaRice.h5.H5FILTER, compression_opts=compression_opts)
+f.close()
+
+f = h5py.File('testFile.h5', 'r')
+test = np.array_equal(f['test'][()], dset)
+f.close()
+assert test, "Failed different M test"
 
 
-if test1 and test2 and test3:
-    print('success!')
-    exit(0)
-else:
-    print('Failure')
-    exit(1)
+f = h5py.File('testFile.h5', 'w')
+compression_opts = (8, 1024)
+f.create_dataset('test', data = dset, compression=deltaRice.h5.H5FILTER, compression_opts=compression_opts)
+f.close()
 
+f = h5py.File('testFile.h5', 'r')
+test = np.array_equal(f['test'][()], dset)
+f.close()
+assert test, "Failed different M segment length test"
+
+
+f = h5py.File('testFile.h5', 'w')
+compression_opts = (8, 1024, 1, 1)
+f.create_dataset('test', data = dset, compression=deltaRice.h5.H5FILTER, compression_opts=compression_opts)
+f.close()
+
+f = h5py.File('testFile.h5', 'r')
+test = np.array_equal(f['test'][()], dset)
+f.close()
+assert test, "Failed different filter test"
+
+
+f = h5py.File('testFile.h5', 'w')
+compression_opts = (8, 1024, 1, 1)
+dset = np.arange(-32768, 32768).astype(np.int16)
+f.create_dataset('test', data = dset, compression=deltaRice.h5.H5FILTER, compression_opts=compression_opts)
+f.close()
+
+f = h5py.File('testFile.h5', 'r')
+test = np.array_equal(f['test'][()], dset)
+f.close()
+assert test, "Failed brute force all possible signed 16-bit values test"
+
+f = h5py.File('testFile.h5', 'w')
+compression_opts = (8, 1024, 1, 1)
+dset = np.arange(0, 65536).astype(np.uint16)
+f.create_dataset('test', data = dset, compression=deltaRice.h5.H5FILTER, compression_opts=compression_opts)
+f.close()
+
+f = h5py.File('testFile.h5', 'r')
[…]
