### 2024-05-10 #6598 @emdupre
Thank you, @david-mathews-1994 ! It looks like these changes to the codebase are associated with these new testing additions :

- https://github.com/david-mathews-1994/deltarice/pull/5
- https://github.com/david-mathews-1994/deltarice/pull/6
- https://github.com/david-mathews-1994/deltarice/commit/a80f7b92765a62673ea551fc8a1392d5eb18ceeb

👋  @macrocosme & @DanNixon & @dineshchitlangia, please let us know whether these effectively changes address your concerns or if you see any remaining issues !

In particular, I noticed that you have not checked the "Community guidelines" criteria, @dineshchitlangia. If you have particular feedback there, please let us know 🙏
