+++ deltaRice.egg-info/requires.txt
@@ -1,0 +1,4 @@
+setuptools>=0.7
+Cython>=0.19
+numpy>=1.6.1
+h5py>=2.4.0
