+++ .github/workflows/wheels.yml
@@ -35,16 +35,14 @@ jobs:
           CIBW_ENVIRONMENT: >
             LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/lib
             CPATH=/usr/local/include
-            ENABLE_ZSTD=1
-            BITSHUFFLE_ARCH=haswell
-          CIBW_TEST_REQUIRES: pytest
+          CIBW_TEST_REQUIRES: python
           # Install different version of HDF5 for unit tests to ensure the
           # wheels are independent of HDF5 installation
           # CIBW_BEFORE_TEST: |
           #   chmod +x .github/workflows/install_hdf5.sh
           #   .github/workflows/install_hdf5.sh 1.8.11
           # Run units tests but disable test_h5plugin.py
-          CIBW_TEST_COMMAND: python pythonExample.py #run the unit tests
+          CIBW_TEST_COMMAND: python {package}/pythonExample.py #run the unit tests
           # The Github runners for macOS don't support AVX2 instructions and so the tests will fail with SIGILL, so skip them
           CIBW_TEST_SKIP: "*macosx*"
 
