### 2024-05-16 b370b86
Update wheels.yml
removing python requirement, that seems to be where it is failing now
