### 2024-05-16 8a45c1c
Update and rename pythonExample.py to tests/pythonExample.py
Updating the test framework to be more thorough.
