### 2024-05-16 8d1e647
doing some cleanup and removing some files
