+++ deltaRice.egg-info/SOURCES.txt
@@ -1,12 +1,0 @@
-LICENSE
-README.md
-setup.py
-deltaRice.egg-info/PKG-INFO
-deltaRice.egg-info/SOURCES.txt
-deltaRice.egg-info/dependency_links.txt
-deltaRice.egg-info/requires.txt
-deltaRice.egg-info/top_level.txt
-src/deltaRice.c
-src/deltaRice_h5plugin.c
-src/h5.c
-src/hdf5_dl.c
