### 2024-05-16 064322b
Update wheels.yml
fixing a few typos
