### 2024-05-16 22c71ea
Update wheels.yml
oops another typo
