+++ .github/workflows/wheels.yml
@@ -42,7 +42,7 @@ jobs:
           #   chmod +x .github/workflows/install_hdf5.sh
           #   .github/workflows/install_hdf5.sh 1.8.11
           # Run units tests but disable test_h5plugin.py
-          CIBW_TEST_COMMAND: pytest {package}/*.py #run the unit tests
+          CIBW_TEST_COMMAND: pytest {package}/tests/*.py #run the unit tests
           # The Github runners for macOS don't support AVX2 instructions and so the tests will fail with SIGILL, so skip them
           CIBW_TEST_SKIP: "*macosx*"
 
