### 2024-05-16 0b3de14
Update wheels.yml
temporarily disabling the macos compilation to see if the ubuntu install works now
