### 2024-05-16 d00f562
Update wheels.yml
making sure that hdf5 is installed before the testing
