### 2024-05-16 fd9667f
Update main.yml
updating how the tests are run in the main workflow
