+++ .github/workflows/main.yml
@@ -56,5 +56,4 @@ jobs:
         python setup.py install --h5plugin --h5plugin-dir ~/hdf5/lib
 
     - name: Run tests
-      run: 
-        python pythonExample.py
+      run: pytest tests/*
