### 2024-05-16 cfa232d
Update wheels.yml
allowing for pypi uploads more easily for debugging
