+++ .github/workflows/wheels.yml
@@ -88,7 +88,7 @@ jobs:
     needs: [build_wheels, build_sdist]
     runs-on: ubuntu-latest
     # Upload to PyPI on every tag
-    if: github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags')
+    # if: github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags')
     # Alternatively, to publish when a GitHub Release is created, use the following rule:
     #if: github.event_name == 'release' && github.event.action == 'published'
     steps:
