### 2024-05-16 1d6264a
Update setup.py
changing capitalization of deltarice to deltaRice in the setup.py file
