### 2024-05-29 #6598 @emdupre
Thanks for checking in, @danielskatz ! I'm still waiting on final reviewer sign-off, in particular re: @dineshchitlangia 's points on the community guidelines criteria.

@dineshchitlangia and @macrocosme, please let us know if you've had a chance to re-review _Delta-Rice_ and are happy with the state of the submission so we can move this forward ! :rocket:
