### 2024-06-05 #6598 @emdupre
@dineshchitlangia Unfortunately this is not yet possible, though it is in the pipeline for future development (see https://github.com/openjournals/joss/issues/1286 ) !

For now, the recommendation would be to add your ORCID to your JOSS reviewer profile (https://reviewers.joss.theoj.org) and make sure that your [GitHub and ORCID are linked](https://info.orcid.org/orcid-and-github-sign-memorandum-of-understanding/). In the future, this should help us to collate your completed reviews. Apologies that there is not a more immediate solution, however.
