### 2024-06-05 5aba5b3
Delete .eggs/Cython-3.0.3-py3.11-linux-x86_64.egg directory
cleaning up the repo some more
