+++ .eggs/Cython-3.0.3-py3.11-linux-x86_64.egg/Cython/Compiler/Symtab.py
@@ -1,2901 +1,0 @@
-#
-#   Symbol Table
-#
-
-from __future__ import absolute_import
-
-import re
-import copy
-import operator
-
-try:
-    import __builtin__ as builtins
-except ImportError:  # Py3
-    import builtins
-
-from ..Utils import try_finally_contextmanager
-from .Errors import warning, error, InternalError
-from .StringEncoding import EncodedString
-from . import Options, Naming
-from . import PyrexTypes
-from .PyrexTypes import py_object_type, unspecified_type
-from .TypeSlots import (
-    pyfunction_signature, pymethod_signature, richcmp_special_methods,
-    get_slot_table, get_property_accessor_signature)
-from . import Future
-
-from . import Code
-
-iso_c99_keywords = {
-    'auto', 'break', 'case', 'char', 'const', 'continue', 'default', 'do',
-    'double', 'else', 'enum', 'extern', 'float', 'for', 'goto', 'if',
-    'int', 'long', 'register', 'return', 'short', 'signed', 'sizeof',
-    'static', 'struct', 'switch', 'typedef', 'union', 'unsigned', 'void',
-    'volatile', 'while',
-    '_Bool', '_Complex'', _Imaginary', 'inline', 'restrict',
-}
-
-
-def c_safe_identifier(cname):
-    # There are some C limitations on struct entry names.
-    if ((cname[:2] == '__' and not (cname.startswith(Naming.pyrex_prefix)
-                                    or cname in ('__weakref__', '__dict__')))
-            or cname in iso_c99_keywords):
-        cname = Naming.pyrex_prefix + cname
-    return cname
-
-def punycodify_name(cname, mangle_with=None):
-    # if passed the mangle_with should be a byte string
-    # modified from  PEP489
-    try:
-        cname.encode('ascii')
-    except UnicodeEncodeError:
-        cname = cname.encode('punycode').replace(b'-', b'_').decode('ascii')
-        if mangle_with:
-            # sometimes it necessary to mangle unicode names alone where
-            # they'll be inserted directly into C, because the punycode
-            # transformation can turn them into invalid identifiers
-            cname = "%s_%s" % (mangle_with, cname)
-        elif cname.startswith(Naming.pyrex_prefix):
-            # a punycode name could also be a valid ascii variable name so
-            # change the prefix to distinguish
-            cname = cname.replace(Naming.pyrex_prefix,
-                                  Naming.pyunicode_identifier_prefix, 1)
-
-    return cname
-
-
-
-
-class BufferAux(object):
-    writable_needed = False
-
-    def __init__(self, buflocal_nd_var, rcbuf_var):
-        self.buflocal_nd_var = buflocal_nd_var
-        self.rcbuf_var = rcbuf_var
-
-    def __repr__(self):
-        return "<BufferAux %r>" % self.__dict__
-
-
-class Entry(object):
-    # A symbol table entry in a Scope or ModuleNamespace.
-    #
-    # name             string     Python name of entity
-    # cname            string     C name of entity
-    # type             PyrexType  Type of entity
-    # doc              string     Doc string
-    # annotation       ExprNode   PEP 484/526 annotation
-    # init             string     Initial value
-    # visibility       'private' or 'public' or 'extern'
-    # is_builtin       boolean    Is an entry in the Python builtins dict
-    # is_cglobal       boolean    Is a C global variable
-    # is_pyglobal      boolean    Is a Python module-level variable
-    #                               or class attribute during
-    #                               class construction
-    # is_member        boolean    Is an assigned class member
-    # is_pyclass_attr  boolean    Is a name in a Python class namespace
-    # is_variable      boolean    Is a variable
[…]
