+++ .eggs/Cython-3.0.3-py3.11-linux-x86_64.egg/Cython/Shadow.py
@@ -1,586 +1,0 @@
-# cython.* namespace for pure mode.
-from __future__ import absolute_import
-
-# Possible version formats: "3.1.0", "3.1.0a1", "3.1.0a1.dev0"
-__version__ = "3.0.3"
-
-try:
-    from __builtin__ import basestring
-except ImportError:
-    basestring = str
-
-
-# BEGIN shameless copy from Cython/minivect/minitypes.py
-
-class _ArrayType(object):
-
-    is_array = True
-    subtypes = ['dtype']
-
-    def __init__(self, dtype, ndim, is_c_contig=False, is_f_contig=False,
-                 inner_contig=False, broadcasting=None):
-        self.dtype = dtype
-        self.ndim = ndim
-        self.is_c_contig = is_c_contig
-        self.is_f_contig = is_f_contig
-        self.inner_contig = inner_contig or is_c_contig or is_f_contig
-        self.broadcasting = broadcasting
-
-    def __repr__(self):
-        axes = [":"] * self.ndim
-        if self.is_c_contig:
-            axes[-1] = "::1"
-        elif self.is_f_contig:
-            axes[0] = "::1"
-
-        return "%s[%s]" % (self.dtype, ", ".join(axes))
-
-
-def index_type(base_type, item):
-    """
-    Support array type creation by slicing, e.g. double[:, :] specifies
-    a 2D strided array of doubles. The syntax is the same as for
-    Cython memoryviews.
-    """
-    class InvalidTypeSpecification(Exception):
-        pass
-
-    def verify_slice(s):
-        if s.start or s.stop or s.step not in (None, 1):
-            raise InvalidTypeSpecification(
-                "Only a step of 1 may be provided to indicate C or "
-                "Fortran contiguity")
-
-    if isinstance(item, tuple):
-        step_idx = None
-        for idx, s in enumerate(item):
-            verify_slice(s)
-            if s.step and (step_idx or idx not in (0, len(item) - 1)):
-                raise InvalidTypeSpecification(
-                    "Step may only be provided once, and only in the "
-                    "first or last dimension.")
-
-            if s.step == 1:
-                step_idx = idx
-
-        return _ArrayType(base_type, len(item),
-                          is_c_contig=step_idx == len(item) - 1,
-                          is_f_contig=step_idx == 0)
-    elif isinstance(item, slice):
-        verify_slice(item)
-        return _ArrayType(base_type, 1, is_c_contig=bool(item.step))
-    else:
-        # int[8] etc.
-        assert int(item) == item  # array size must be a plain integer
-        return array(base_type, item)
-
-# END shameless copy
-
-
-compiled = False
-
-_Unspecified = object()
-
-# Function decorators
-
-def _empty_decorator(x):
-    return x
-
-def locals(**arg_types):
-    return _empty_decorator
-
-def test_assert_path_exists(*paths):
-    return _empty_decorator
-
-def test_fail_if_path_exists(*paths):
-    return _empty_decorator
-
-class _EmptyDecoratorAndManager(object):
-    def __call__(self, x):
-        return x
-    def __enter__(self):
-        pass
-    def __exit__(self, exc_type, exc_value, traceback):
-        pass
-
-class _Optimization(object):
-    pass
-
-cclass = ccall = cfunc = _EmptyDecoratorAndManager()
-
-annotation_typing = returns = wraparound = boundscheck = initializedcheck = \
-    nonecheck = embedsignature = cdivision = cdivision_warnings = \
-    always_allows_keywords = profile = linetrace = infer_types = \
-    unraisable_tracebacks = freelist = \
-        lambda _: _EmptyDecoratorAndManager()
-
[…]

[…]
