### 2024-06-05 #6598 @david-mathews-1994
@emdupre Hello! I've linked the repo with Zenodo and have the following DOI for it: https://doi.org/10.5281/zenodo.11490673. It doesn't seem to link properly yet, I'm hoping that's just a bit of a delay since I just put that out there. The version tag is v1.0.0.
