### 2024-06-06 #6598 @emdupre
Thank you, @david-mathews-1994 ! I do have a few editorial requests :

## On the repository:
- [ ] Could you please fix the packaging such that tests pass ? [Currently tests fail](https://github.com/david-mathews-1994/deltarice/actions/runs/9387550194/job/25850729166) as `deltarice` is not installed in the testing environment. It looks like this change was introduced in 1f6a280.

## On the software paper:
- [ ] Could you confirm that authors H. Li and R. Mammei do not have ORCIDs ?
- [ ] Could you please remove unused references from the associated bib file ? These are being checked by editorial bot and introducing a few unnecessary warnings.
- [ ] Similarly, could you confirm that `OldPaper.md` is still intended to be in the `paper` directory ?
- [ ] I'm somewhat confused about the location of footnote [1]. I believe that this text is better placed in the Acknowledgements rather than as a footnote. I have consulted with other members of the editorial board on this, however, and will report back.
- [ ] The text contains references to both "G-zip" (L20) and "Gzip" (L34) ; could you please standardize these ? I believe that "Gzip" is correct. Could you please also add a citation for this software.
- [ ] On L36, please ensure that their is a single space between text and supporting citations.
- [ ] Could you please add a title and caption to Table 1? On the same table, could you confirm that the "Output" row is intended to be blank ?
- [ ] On L99 (in the Fry et al. reference) I believe that Nab should be capitalized given its notation in-text. Could you please adjust this, if so ?

Also, apologies, but I noticed that you're using the AAS collaboration template. This wasn't flagged during the pre-review, so I just want to confirm : is this manuscript a companion to an AAS submission ([as described here](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing)) ?
