### 2024-06-11 #6598 @david-mathews-1994
@emdupre Apologies for the delay on this, I was out of the office for a bit.

I can fix the tests. I was in the process of trying to make the software installable via pip and was working on a automated wheels deployment but haven't quite gotten that to work correctly. I'll hopefully have that fixed later today.

I've reached out to those collaborators and am waiting on a reply from them currently.

Regarding footnote [1], that is text that is required from my employer for acknowledgement of funding/resources. I've received push-back from them in the past when it wasn't on the first page for other journal submissions, so if possible I'd like to keep it there. If that isn't possible, then I can just tell them that it isn't allowed by the journal and I think that'll be okay so long as it remains in the document.

Regarding the template, this is not an AAS submission. I was simply following the example here  [https://joss.readthedocs.io/en/latest/example_paper.html] and didn't modify that part of the document. What would you recommend I change that to?
