### 2024-06-11 b9527d6
Delete paper/OldPaper.md
