### 2024-06-11 fa63079
Update paper.md
fixing some formatting issues
