### 2024-06-11 #6598 @emdupre
Thanks for following up ! Please see inline responses below.

> I can fix the tests. I was in the process of trying to make the software installable via pip and was working on a automated wheels deployment but haven't quite gotten that to work correctly. I'll hopefully have that fixed later today.

> I've reached out to those collaborators and am waiting on a reply from them currently.

Great, thank you ! Please let me know when you have updates on these points.

> Regarding footnote [1], that is text that is required from my employer for acknowledgement of funding/resources. I've received push-back from them in the past when it wasn't on the first page for other journal submissions, so if possible I'd like to keep it there. If that isn't possible, then I can just tell them that it isn't allowed by the journal and I think that'll be okay so long as it remains in the document.

I reached out to other members of the editorial team on this ; they noted that the terms listed in the notice are already compatible with JOSS policies because authors retain copyright and CC-BY (the license under which the paper text is release) is entirely compatible with the terms above.
I'll also note that since we generate the final PDF directly from the software repository, there isn't a separate step to remove this notice for the version of record.
Given that, do you still want to include it in the manuscript ? If so, I think the acknowledgements is the right place for it. Otherwise, you can also point back to this (publicly available) review comment thread to note that JOSS acknowledges the statement.

> Regarding the template, this is not an AAS submission. I was simply following the example here [https://joss.readthedocs.io/en/latest/example_paper.html] and didn't modify that part of the document. What would you recommend I change that to?

You should remove the associated optional fields if this is not an AAS associated submission ; specifically, these lines :
```
# Optional fields if submitting to a AAS journal too, see this blog post:
# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
aas-journal: Astrophysical Journal <- The name of the AAS journal.
```
