### 2024-06-11 f2523c6
Update paper.md
cleaning up footnotes and the ORNL acknowledgement
