### 2024-06-11 4e57051
Update setup.py
removing the packages=["deltarice"] flag as that may be causing the build issues
