### 2024-06-11 f9402d5
Update test.py
adding back the rest of the testing
