### 2024-06-11 #6598 @david-mathews-1994
So in speaking with others at my work, it doesn't sound like we are required to post the footnote in the published version of the document so long as the publication acknowledges the requirements and I believe your last message satisfies that requirement which is great! I've gone ahead and removed it from the document. I also removed the AAS lines and the newest version of the PDF should reflect these changes.

I also fixed the tests failing issues and all are passing now. Just forgot to remove a line in the setup.py script that was breaking things from a previous push (this is why tests are great though!).
