### 2024-06-13 3618610
Update paper.bib
removing unnecessary references
