+++ paper/paper.bib
@@ -11,17 +11,6 @@ keywords = {Parity violation, Scintillator detector, Neutron resonance, Fundamen
 abstract = {We describe a modular apparatus for use in parity-violation measurements in epithermal neutron–nucleus resonances with high instantaneous neutron fluxes at the Manuel Lujan Jr. Neutron Scattering Center at Los Alamos National Laboratory. This apparatus is designed to conduct high-precision measurements of the parity-odd transmission asymmetry of longitudinally polarized neutrons through targets containing nuclei with p-wave neutron–nucleus resonances in the 0.1–10 eV energy regime and to accommodate a future search for time reversal violation in polarized neutron transmission through polarized nuclear targets. The apparatus consists of an adjustable neutron and gamma collimation system, a 3He-4He ion chamber neutron flux monitor, two identical cryostats for target cooling, an adiabatic eV-neutron spin flipper, a near-unit efficiency 6Li-7Li scintillation detector operated in current mode, a flexible CAEN data acquisition system, and a neutron spin filter based on spin-exchange optical pumping of 3He gas. We describe the features of the apparatus design devoted to the suppression of systematic errors in parity-odd asymmetry measurements. We describe the configuration of the apparatus used to conduct a precision measurement of parity violation at the 0.7 eV p-wave resonance in 139La which employs two identical 139La targets, one to polarize the beam on the p-wave resonance using the weak interaction and one to analyze the polarization.}
 }    %url = {https://www.sciencedirect.com/science/article/pii/S0168900220304307},
 
-@ARTICLE{huffman,
-    author={Huffman, David A.},
-    journal={Proceedings of the IRE}, 
-    title={A Method for the Construction of Minimum-Redundancy Codes}, 
-    year={1952},
-    volume={40},
-    number={9},
-    pages={1098-1101},
-    doi={10.1109/JRPROC.1952.273898}
-}
-
 @ARTICLE{golomb,
     author={Golomb, S.},
     journal={IEEE Transactions on Information Theory}, 
@@ -42,18 +31,6 @@ abstract = {We describe a modular apparatus for use in parity-violation measurem
   number={6},
   pages={889-897},
   doi={10.1109/TCOM.1971.1090789}}
-
-
-@article{nab,
-    author = {Broussard, Leah and Baeßler, S. and Bailey, Thomas and Birge, Noah and Bowman, David and Crawford, Christopher and Cude-Woods, C. and Fellers, D. and Fomin, N. and Frlez, Emil and Gericke, M. and Hayen, L. and Jezghani, A. and Li, H. and Macsai, N. and Makela, Mark and Mammei, Russell and Mathews, D. and McGaughey, P. and Zeck, B.},
-    year = {2018},
-    month = {12},
-    pages = {},
-    title = {Using Nab to determine correlations in unpolarized neutron decay},
-    volume = {240},
-    journal = {Hyperfine Interactions},
-    doi = {10.1007/s10751-018-1538-7}
-}
 
 @article{MASUI2015181,
     title = {A compression scheme for radio data in high performance computing},
@@ -100,75 +77,6 @@ month = {2},
   journal = {{EPJ} Web of Conferences}
 }%url = {https://doi.org/10.1051/epjconf/201921904002},
 
-@article{Ahmed_2019,
-    doi = {10.1088/1748-0221/14/11/P11017},
-    year = {2019},
-    month = {nov},
-    publisher = {},
-    volume = {14},
-    number = {11},
-    pages = {P11017},
[…]
