### 2024-06-13 b24d792
Update paper.md
adding in reference to david's thesis
