### 2024-06-13 b8e2d1a
Update paper.md
adding in russ's ORCID
