### 2024-06-13 #6598 @david-mathews-1994
I've reached out to my collaborators and found the two missing ORCID numbers so that is taken care of. I believe I've also standardized the references to Gzip at this time.

Regarding the table, how do I actually put in a caption for that table within Markdown or in such a way that the Latex interpreter system understands it? I know how to do this in pure Latex, but am not sure how to add a caption to a non-figure object in a way that this PDF generator understands
