### 2024-06-17 #6598 @emdupre
Thank you, @david-mathews-1994 !

The [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures) _should_  cover this point, but it looks like that section has a formatting issue ; I'll submit an issue on this, so thanks for flagging !

In general, this can be done as simply as adding the following line before the table :
```
Table 1: An example of Rice coding when writing $-2, 25$ with $m=8$
```

You can add a label as well, as discussed in the documentation, s.t. it can be directly referenced in text, replacing "in the table below."
