### 2024-06-17 ffeb908
Update paper.md
updating the bit packing table
