+++ paper/paper.md
@@ -108,14 +108,16 @@ Many modern nuclear physics experiments, such as the Nab experiment, will produc
 This algorithm is a two-step process: the digitized signal is first passed through an encoding operation, such as delta encoding, to de-correlate the data and prepare it for the second step of Rice coding [@rice]. These methods were chosen for this compression routine specifically for their simplicity, throughput, and storage efficiency. They also do not require a significant amount of additional information to be stored alongside the compressed data in order for the decompression routine to function, which improves storage efficiency further. 
 
 ## Rice Coding 
-Rice coding functions by encoding a value $x$ in 2 pieces: $q$, the result of a division by a tunable parameter $m$, and $r$, the remainder of that division. $q$ is stored in Unary coding, with $r$ in truncated binary. In this routine, signed values are handled by interleaving positive and negative values as follows: $x\prime = 2*x$ for $x>=0$ and $x\prime = 2|x|-1$ for $x<0$. Rice coding is used instead of the more general Golomb coding [@golomb] because the restriction to powers of $2$ for $m$ allows for more efficient calculations. For information about the optimization of $m$, see [Optimization](https://github.com/david-mathews-1994/deltarice/blob/master/docs/Optimization.md). In the case that $q>=8$, the output will be $q=8$ followed by the original number in 16-bit signed representation. This is done to ensure that the amount a value can fail to be compressed is fixed. The outputs from this method are packed sequentially into 32 bit containers ensuring that no bits are wasted for any containers but the last one for a dataset. The table below shows an example of this bit packing when writing $-2, 25$ with $m=8$. In this example, the output is formatted as 8 bit packets, and a temporary 16 bit storage location is used to do the packing. 
+Rice coding functions by encoding a value $x$ in 2 pieces: $q$, the result of a division by a tunable parameter $m$, and $r$, the remainder of that division. $q$ is stored in Unary coding, with $r$ in truncated binary. In this routine, signed values are handled by interleaving positive and negative values as follows: $x\prime = 2*x$ for $x>=0$ and $x\prime = 2|x|-1$ for $x<0$. Rice coding is used instead of the more general Golomb coding [@golomb] because the restriction to powers of $2$ for $m$ allows for more efficient calculations. For information about the optimization of $m$, see [Optimization](https://github.com/david-mathews-1994/deltarice/blob/master/docs/Optimization.md). In the case that $q>=8$, the output will be $q=8$ followed by the original number in 16-bit signed representation. This is done to ensure that the amount a value can fail to be compressed is fixed. The outputs from this method are packed sequentially into 32 bit containers ensuring that no bits are wasted for any containers but the last one for a dataset. Table \ref{table:bitpacking} shows an example of bit packing for 8 bit output packets with a 16 bit temporary storage location.
 
-| Bit          | 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14 | 15 |
[…]
