### 2024-06-17 55adc45
Update paper.md
adjusting table some more
