### 2024-06-17 98c6019
Update paper.md
trying again to fix the table
