### 2024-06-17 83aaa97
Update paper.md
adjusting table some more
