### 2024-06-19 266505f
Update paper.md
adding in the table as a figure
