### 2024-06-19 ba54d2e
Update paper.md
fixing the caption for the table
