### 2024-06-19 #6598 @david-mathews-1994
@emdupre I ran into a bunch of issues trying to get the table to work so I ended up just making the table in external software as an image and put it into the document that way. Let me know if that formatting looks better!
