### 2024-06-20 #6598 @emdupre
Thank you, @david-mathews-1994 !

I realized I overlooked formatting request, apologies : Currently, affiliations include just the relevant institute. Could you please list each affiliation in a "Department, Institute/Univ., City, Country" format ? Or as close to that format as possible, recognizing that some institutes do not have departments.
