### 2024-06-20 #6598 @emdupre
Thank you for those updates !

The remaining steps that I see :
- There is a [remaining checklist item](https://github.com/openjournals/joss-reviews/issues/6598#issuecomment-2151248287) for formatting the capitalization in the references
- There is a [remaining post-review author step](https://github.com/openjournals/joss-reviews/issues/6598#issuecomment-2145543940) to make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. Currently, the title, author listing, and license are all non-overlapping with the code repository. These can be updated by editting the metadata for the Zenodo archive, as discussed here : https://help.zenodo.org/docs/deposit/manage-records/#edit
  This should _not_ require creating a new version, but please let me know if you hit any difficulties !
