### 2024-06-21 5c929e6
Update paper.bib
trying to fix capitalization
