+++ paper/paper.bib
@@ -73,7 +73,7 @@ month = {2},
   pages = {04002},
   author = {J. Fry and R. Alarcon and S. Bae{\ss}ler and S. Balascuta and L. Barr{\'{o}}n Palos and T. Bailey and K. Bass and N. Birge and A. Blose and D. Borissenko and J.D. Bowman and L.J. Broussard and A.T. Bryant and J. Byrne and J.R. Calarco and J. Caylor and K. Chang and T. Chupp and T.V. Cianciolo and C. Crawford and X. Ding and M. Doyle and W. Fan and W. Farrar and N. Fomin and E. Frle{\v{z}} and M.T. Gericke and M. Gervais and F. Gl\"{u}ck and G.L. Greene and R.K. Grzywacz and V. Gudkov and J. Hamblen and C. Hayes and C. Hendrus and T. Ito and A. Jezghani and H. Li and M. Makela and N. Macsai and J. Mammei and R. Mammei and M. Martinez and D.G. Matthews and M. McCrea and P. McGaughey and C.D. McLaughlin and P. Mueller and D. van Petten and S.I. Penttil\"{a} and D.E. Perryman and R. Picker and J. Pierce and D. Po{\v{c}}ani{\'{c}} and Y. Qian and J. Ramsey and G. Randall and G. Riley and K.P. Rykaczewski and A. Salas-Bacci and S. Samiei and E.M. Scott and T. Shelton and S.K. Sjue and A. Smith and E. Smith and E. Stevens and J. Wexler and R. Whitehead and W.S. Wilburn and A. Young and B. Zeck},
   editor = {T. Jenke and S. Degenkolb and P. Geltenbort and M. Jentschel and V.V. Nesvizhevsky and D. Rebreyend and S. Roccia and T. Soldner and A. Stutz and O. Zimmer},
-  title = {The Nab experiment: A precision measurement of unpolarized neutron beta decay},
+  title = {The {Nab} experiment: A precision measurement of unpolarized neutron beta decay},
   journal = {{EPJ} Web of Conferences}
 }%url = {https://doi.org/10.1051/epjconf/201921904002},
 
