### 2024-06-21 #6598 @david-mathews-1994
@emdupre The capitalization in the references has now been fixed. I also went through and updated the Zenodo archive to match the author list and verified the ORCIDs are all matching now as well
