### 2024-06-21 #6598 @emdupre
Thank you, @david-mathews-1994 ! The archive and software paper now look correct to me.

I'm now happy to recommend _Delta-Rice_ to the EiC team for publication in JOSS ✨🚀

Thank you, too, to @macrocosme, @DanNixon, and @dineshchitlangia for your reviews ! JOSS works because of your work 💐
