### 2024-06-21 #6598 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
