### 2024-06-21 33a8b4a
minor changes in paper.md
