+++ paper/paper.md
@@ -82,9 +82,9 @@ affiliations:
    index: 7
  - name: Georgia Institute of Technology, Atlanta, GA, USA
    index: 8
- - name: University of Manitoba, Winnipeg, CA
+ - name: University of Manitoba, Winnipeg, Canada
    index: 9
- - name: University of Winnipeg, Winnipeg, CA
+ - name: University of Winnipeg, Winnipeg, Canada
    index: 10
  - name: Arizona State University, Tempe, AZ, USA
    index: 11
@@ -113,7 +113,7 @@ Rice coding functions by encoding a value $x$ in 2 pieces: $q$, the result of a 
 ![A demonstration of rice coding and bit packing when writing $x=-2$ and $x=25$ with $m=8$ for a $8$ bit output container with a 16 bit temporary cache. Any remaining data in the temporary buffer is retained for the next write of $x$, or output at the end of the compression when no more values of $x$ are provided.](./../docs/images/BitPackingTable.png?raw=true)
 
 ## Preparatory Encoding
-Preparatory encoding is done to adjust the dataset to a form more optimal for Rice Coding. By default, this is done with delta encoding which stores the difference between subsequent values. The image below shows an example of this when applied to a signal from the Nab experiment. A simple optimization routine for determining the ideal filter is discussed in [Optimization](https://github.com/david-mathews-1994/deltarice/blob/master/docs/Optimization.md). 
+Preparatory encoding is done to adjust the dataset to a form more optimal for Rice Coding. By default, this is done with delta encoding, which stores the difference between subsequent values. The image below shows an example of this when applied to a signal from the Nab experiment. A simple optimization routine for determining the ideal filter is discussed in [Optimization](https://github.com/david-mathews-1994/deltarice/blob/master/docs/Optimization.md). 
 
 ![Left: A waveform before and after delta encoding. Applying Rice coding with $m=8$ on the original signal expands the size of the waveform from 14 kB to 18.2 kB. The same Rice coding operation on the delta encoded waveform compresses the waveform to 4.6 kB, 33% the original size. Right: A histogram of a sample dataset before and after delta encoding. Note the clear reduction in the distribution width and that the most probable values are centered around 0.](./../docs/images/ExampleEncoding.png?raw=true)
 
@@ -123,7 +123,7 @@ Delta-Rice is accessible to users through the HDF5 library [@hdf5] as filter ID 
 
 # Acknowledgements
 
-This research was sponsored by the U.S. Department of Energy (DOE), Office of Science, Office of Nuclear Physics [contracts DE-AC05-00OR22725, DE-SC0014622, DE-FG02-03ER41258] and National Science Foundation (NSF) [contract PHY-1812367]. This research was also sponsored by the U.S. Department of Energy, Office of Science, Office of Workforce Development for Teachers and Scientists (WDTS) Graduate Student Research (SCGSR) program. This research was supported in part through research cyberinfrastructure resources and services provided by the Partnership for an Advanced Computing Environment (PACE) at the Georgia Institute of Technology, Atlanta, Georgia, USA.
+This research was sponsored by the U.S. Department of Energy (DOE), Office of Science, Office of Nuclear Physics [contracts DE-AC05-00OR22725, DE-SC0014622, DE-FG02-03ER41258] and National Science Foundation (NSF) [award PHY-1812367]. This research was also sponsored by the U.S. Department of Energy, Office of Science, Office of Workforce Development for Teachers and Scientists (WDTS) Graduate Student Research (SCGSR) program. This research was supported in part through research cyberinfrastructure resources and services provided by the Partnership for an Advanced Computing Environment (PACE) at the Georgia Institute of Technology, Atlanta, Georgia, USA.
 
 # References
 
