### 2024-06-21 #6598 @danielskatz
@david-mathews-1994 - the few needed changes I found are in https://github.com/david-mathews-1994/deltarice/pull/9

Please merge this, or let me know what you disagree with, and then we can finish the acceptance and publication of this work.
